Address review comments: add useMemo to handleWalletLogin hook, rename Loading component + other minor fixes
